src/servers/version_control/requirements.txt
uvicorn==0.34.0
fastapi==0.115.12
websockets==15.0.1
aiofiles==24.1.0
httpx==0.28.1
python-dotenv==1.1.0
pydantic==2.11.2